refactor: extract components to make a/b design implementation easier
