+++ src/app/(registration)/sign-up/[[...sign-up]]/page.tsx
@@ -2,42 +2,17 @@
 "use client";
 
 import { SignUp } from "@clerk/nextjs";
-import {
-  OakBox,
-  OakFlex,
-  OakHeading,
-  OakIcon,
-  OakInlineBanner,
-  OakLI,
-  OakUL,
-} from "@oaknational/oak-components";
-import { PropsWithChildren } from "react";
+import { OakInlineBanner } from "@oaknational/oak-components";
 
 import { formAppearanceStyles } from "../../formAppearanceStyles";
 import { AuthLayout } from "../../auth-layout";
 
-import { getIllustrationAsset } from "@/image-data";
-import CMSImage from "@/components/SharedComponents/CMSImage";
-
-function ListItem({ children }: PropsWithChildren) {
-  return (
-    <OakLI
-      $display="flex"
-      $flexDirection="row"
-      $alignItems="center"
-      $gap="space-between-s"
-      $mb="space-between-xs"
-    >
-      <OakIcon iconName="tick" $height="all-spacing-7" $width="all-spacing-7" />
-      {children}
-    </OakLI>
-  );
-}
+import RegistrationAside from "@/components/TeacherComponents/RegistrationAside/ResgistrationAside";
 
 function SignUpPage() {
   return (
     <AuthLayout
-      headerSlot={
+      bannerSlot={
         <OakInlineBanner
           isOpen
           message={
@@ -49,33 +24,7 @@ function SignUpPage() {
           $mb={["space-between-none", "space-between-m"]}
         />
       }
-      asideSlot={
-        <OakBox
-          $width="min-content"
-          $minWidth="all-spacing-21"
-          $mb="space-between-xl"
-        >
-          <OakFlex $mb="space-between-m2" $maxHeight="all-spacing-19">
-            <CMSImage
-              image={getIllustrationAsset("auth-acorn")}
-              $objectFit="contain"
-            />
-          </OakFlex>
-          <OakHeading
-            tag="h1"
-            $font="heading-5"
-            $mb="space-between-l"
-            $textAlign="center"
-          >
-            Our resources will always be free. Creating an account gives you:
-          </OakHeading>
-          <OakUL $font="heading-light-7" $reset $mh="space-between-xl">
-            <ListItem>Full unit downloads</ListItem>
-            <ListItem>Instant access to Aila, our AI lesson assistant</ListItem>
-            <ListItem>Priority access to new products and features</ListItem>
-          </OakUL>
-        </OakBox>
-      }
+      asideSlot={<RegistrationAside />}
     >
       <SignUp
         appearance={formAppearanceStyles}
